Removed mxnet test_utils.py dependency on nose.
