+++ tests/python/gpu/test_operator_gpu.py
@@ -22,11 +22,12 @@ import unittest
 import mxnet as mx
 import numpy as np
 import unittest
-from mxnet.test_utils import *
+from mxnet.test_utils import check_consistency, set_default_context, assert_almost_equal
 from numpy.testing import assert_allclose
 
 curr_path = os.path.dirname(os.path.abspath(os.path.expanduser(__file__)))
 sys.path.insert(0, os.path.join(curr_path, '../unittest'))
+from common import *
 from test_operator import *
 from test_optimizer import *
 from test_random import *
