+++ tests/python/unittest/test_executor.py
@@ -17,7 +17,7 @@
 
 import numpy as np
 import mxnet as mx
-from mxnet.test_utils import *
+from common import *
 
 
 def reldiff(a, b):
