+++ tests/python/unittest/test_gluon.py
@@ -18,7 +18,8 @@
 import mxnet as mx
 from mxnet import gluon
 from mxnet.gluon import nn
-from mxnet.test_utils import *
+from mxnet.test_utils import assert_almost_equal
+from common import *
 import numpy as np
 from nose.tools import raises
 from copy import deepcopy
