+++ tests/python/unittest/test_random.py
@@ -18,7 +18,7 @@
 import os
 import mxnet as mx
 import numpy as np
-from mxnet.test_utils import *
+from common import *
 
 def same(a, b):
     return np.sum(a != b) == 0
